+++ lib/src/v2/substituter.dart
@@ -2,6 +2,7 @@ import 'package:postgres/src/types/generic_type.dart';
 
 import '../types.dart';
 
+@Deprecated('Do not use v2 API, will be removed in next release.')
 class PostgreSQLFormat {
   static String id(String name, {Type? type}) {
     if (type != null) {