+++ lib/src/v2/transaction_proxy.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: deprecated_member_use_from_same_package
+
 part of 'connection.dart';
 
 /// Represents a rollback from a transaction.
@@ -5,6 +7,7 @@ part of 'connection.dart';
 /// If a transaction is cancelled using [PostgreSQLExecutionContext.cancelTransaction], the value of the [Future]
 /// returned from [PostgreSQLConnection.transaction] will be an instance of this type. [reason] will be the [String]
 /// value of the optional argument to [PostgreSQLExecutionContext.cancelTransaction].
+@Deprecated('Do not use v2 API, will be removed in next release.')
 class PostgreSQLRollback {
   @internal
   PostgreSQLRollback(this.reason);